+++ src/main/java/org/jenkinsci/plugins/sonargerrit/review/GerritReviewBuilder.java
@@ -117,6 +117,7 @@ public class GerritReviewBuilder {
     commentInput.id = input.getKey();
     commentInput.line = input.getLine();
     commentInput.message = message;
+    commentInput.unresolved = true;
     return commentInput;
   }